Get method to search for a game from Rawg

+++ src/main/java/com/example/backloggd/Repository/GameRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GameRepository extends JpaRepository<GamesModel, Long> {
+public interface GameRepository extends JpaRepository<GamesModel, Integer> {
    Optional<GamesModel> findByGameName(String gameName);
 }
